Refactor filters crate into modular components
